Make the MuonStationIntersectSvc more thread-safe
The service now uses the MdtCondDbData object created by the MdtCondDbAlg, rather than the MdtCondSummarySvc.
The MdtIntersectGeometry objects are no longer cached in the service but created on the fly as needed, making the service much simpler.
Mutable data members are removed or made local variables.
The MdtCondDbData object and the MuonDetectorManager must now be passed in from outside: the clients of the service are adjusted accordingly.
The MdtCondDbAlg must now run during muon reconstruction: the configuration is updated accordingly.