Fixed the CMake configuration of TrackToCalo.
This was *very* painful. :-( The shared library of the package ends up
including all the symbols of the components that should in principle
only make it into the component library of the package. But that's
not how things are set up.

So many of the private dependencies of the shared library have to be
repeated as dependencies of the component library as well. Otherwise
the entries.cxx file would have issues with including the "private"
headers from the src/ directory.

In principle the package should be re-arranged quite a bit to clean
things...