Optimise DAOD_PHYS tau content
Hello,

This MR implements tau jet and tau track thinning in the DAOD_PHYS format, as discussed in ATLASG-1502.
The TauTrackParticleThinning code has been extended to thin tau tracks, in addition to ID tracks.
To be consistent with the existing code, a functionality to thin tracks based on dR(tau,track) has been added to TracksInCone.h.
But this is for consistency purpose, it is not recommended to select tau tracks based on dR.
One should rather rely the tau track classification flag.
In this MR, we also drop 'bdtScores' from the tau track smart slimming list, as it is rather big, and not used by analyses as far as we know.
Therefore, it is expected that this MR changes the derivation output for formats using TauJets as a SmartCollection.

Cheers,
Bertrand