Add temporary analyses in a futile attempt to add some ID monitoring for MT
Because some fool changed all the ID Track collection names without
actually consulting anyone, and didn;t really know what they were doing,
the monitoring broke. In addition the idperf chains upon which we rely
have not been implemented.

This attempts to add some monitoring for standard chains, using the new
track cllections. The new collections may not work, it may be that code
changes are needed to use the new TDT. If the monitoring does not work
because of the actual names themselves, then this won't be fixed since
we are in any case moving to a new naming convention, so will be fixed
during the Great Renaming.