Add invariant mass muon trigger hypo
Adding a hypo alg/tool to cut on the invariant mass of dimuon pairs within a mass window (considers all dimuon pairs in the event, and passes if at least one is within the mass window).
Setting up the view links in the combiner hypo, since the invariant mass hypo expects them to be set (and it runs after the combiner hypo).
Also added a preliminary menu sequence (ATR-20126,ATR-20049).