AthenaPoolCnvSvc: Pass SG key to createPersistent/createTransient.
Rename T_AthenaPoolCustomCnv to T_AthenaPoolCustomCnvWithKey,
changing the interfaces for the create methods to pass through the SG key.
For compatibility, create T_AthenaPoolCustomCnv deriving from it
with the old interfaces.

Update T_AthenaPoolxAODCnv and T_AthenaPoolAuxContainerCnv to use this.
In the process, we can remove the xAODCnv hack where we override
createObj() in order to get the SG key.  (This turns out to be a race;
see ATEAM-567.)

A couple other minor cleanups.