Protect against negative elements on the diagonal of covariance matrices
In rare occasions, the track fit may result in parameters with unphysical covariance matrices.
This can occur if the inputs to the fit are themselves unphysical: e.g., ID tracks with pT of hundreds of TeV.
This fix adds a protection against such tracks, to avoid crashes down the line.