Update decoding config in new JO for trigger
Brining the configuration of the decoding for the L2 SA muons in the new config inline with what we use in the current configuration (ie, switch off the calls to the decoding tools from inside the L2 algorithms, since we call the offline tools directly).
Also enable the CSC cluster building for the trigger.