Change the muon data decoding test with caching so it is in INFO mode, not DEBUG.
I also removed a DoSeededDecoding setting for the RPC raw data decoding
in trigger mode. The exact way we configure for the trigger needs
to be sorted out once we actually try to use the new JO fragments
in the new athenaMT setup. For now in the cache test, we just decode the full
detector.
The reference file for the test is also updated.