Update the trigger newJO test to run on 20 events
Changing the number of events we run the new JO test to run over from 10 to 20. This is the same number of events we run, for example, teh full menu test. Since there are no muon RoIs in the first ten events, we couldn't reliably catch if the muon trigger part of the test broke down (20 events should help here).