Remove check for JetFitter_isDefaults for TrackRelativeEta
It seems like this check won't give us consistent results with the current
release. It's not totally clear (without a lot more head scratching) how we
should be setting the default values for the TrackRelativeEta variable.
Right now they are NAN some of the time but it's unclear if this is tied
to one of the _isDefault variables.