Add TauJet EDM protection against invalid TauTrack links (thinned tracks)
Hello,

This MR adds a protection against invalid TauTrack links when retrieving the tau tracks or the number of tau tracks.
This is required to support tau track thinning, now used in DAOD_PHYS (ATLASG-1502).
Now, only the tau tracks that have a valid link are considered.
In particular, the behaviour of nAllTracks() had to be changed to be consistent with allTracks().size().
If one only keeps classifiedCharged tau tracks by thinning away the other tracks, nAllTracks() will now return the number of classifiedCharged tracks.

Cheers,
Bertrand