AthenaPoolUtilities: Don't consider a new xAOD variable a test failure.
To verify that we can properly read old data files, we dump them out
and compare with reference files.

But if a new xAOD variable is added to a class, then this new variable
will also appear in the dump files, leading to test failures.
However, this is really beside the point of what we're trying to test here.

Tweak the reference file comparison so that the tests will not fail just
because a new xAOD variable was added (though a warning will be printed
in that case).