Add straight-line propagation to MooTrackFitter
Currently, the MooTrackFitter code always attempts to perform no-field straight
line propagation. I have concluded that this is not actually correct, and doing
straight-line propagation by default decreases the physics performance in
reconstruction.

This bug has laid dormant for ages because the STEP propagator which is used is
unable to actually do straight-line propagation, and it will always use a
full-field propagation. This means that even if the MooTrackFitter is
configured to do straight-line propagation, the STEP propagator is actually
doing full-field propagation, hiding the error.

This commit fixes the aforementioned error by adding an additional config
variable SLProp (with sensible defaults) so the MooTrackFitter can be
configured in any way that is desirable.