athenaHLT: Do not set concurrency flags in python
Setting the concurrency flags in athenaHLT (threads/events/slots) is not
a good idea as this deviates from what would happen in a partition and
prevents us from reproducing certain bugs (e.g. ATR-20362).

If we had to set these flags it should be done in TrigPSCPythonSetup.py
as this is used both in athenaHLT and in a partition (when running from
job options). However, setting these flags should not be necessary as we
correctly set these properties on the C++ side from values in OKS.

The only downside is that there are certain printouts in the log file
that now show (confusingly) e.g. 0 threads during the python configuration.
On the other side, this may help discovering places where we have a
dependency on these values.