TestHepMC: Tidying and adding a check for valid PDG IDs
The primary purpose of this commit is to add a check for invalid PDG
IDs, stemming from the discussion in !26148 .  This test is optional,
though on by default, and the user can provide a white list of IDs that
should be made exceptions to the rule, to allow exotic PDG IDs that
might not conform to still run.

At the same time, tweaking some of the logic in the filter to set a
boolean when the filter fails, then check the boolean at the end of
execution to push the filter results. This ensures that all tests run,
and that multiple tests can fail on a single event if appropriate. I
think this will be a help when debugging, but it does result in a change
of logging output (of course, the rate at which the filter fails is
still the same!).

The last part is general tidying, white space cleaning, and so on. Watch
for the "ignore white space changes" button in the diff to get to the
interesting part. This removes tabs in favor of spaces, makes the
spacing consistent, ensures we don't have multiple commands on a single
line, and so on. I tried to add a lot of comments as well, so that it's
a bit easier to see where one part of the code is ending and another is
beginning. I hope it's useful!