AthenaKernel: ExtendedEventContext dynamic loader workarounds
Move the helpers for finding MT thinning decisions from DataModelAthenaPool
to AthenaKernel.

As a result, I was seeing again bad_any_cast exceptions, because an internal
symbol used for any_cast was resolved to different addresses in different
libraries, causing any_cast to spuriously fail.
Try to deal with this by introducing a set of out-of-line helper
functions for accessing the extended event context that we use
in preference to directly using the templated methods of EventContext.