Update configuration for full scan muon trigger hypo tools
For muon triggers running full scan muon finding, we want to make sure that the full scan muon hypo considers the number/thresholds for all muons in the full chain (to avoid double counting muons that were all ready found in RoIs).
The hypo tool configuration was updated to take the thresholds from the full chain name (instead of just the full scan part). Also added in a FS RoI building to the muon event view creator. This can be revisted/removed once the official version of the serial merging is available (the very simple version currently in use for these chains has some limitations that make this necessary in the intermediate term to avoid running the full scan algorithms multiple times based on the number of initial L1 RoIs). This is part of the mgration in ATR-17553.