StoreGate: WDHK dependency fix.
WriteDecorHandleKey is meant to be compatible with a WriteHandleKey for the
same container.  It had code to suppress a circular dependency error if the
two keys were used by the same component, but we could still get a warning
if two such keys were used by different components under the same algorithm.

Extend this to work across all components of an algorithm, but try to only
suppress the warning if the component with the WriteHandleKey
is encountered before the component with the WriteDecorHandleKey.