Removing extra copy of event weights
There are currently two identical copies of the event weights stored for
each event: one as a part of the TruthEvent, and one as a part of the
EventInfo. The official weight access tool uses the one from the
EventInfo, so we should keep that one and remove the other.

Total savings: 0.7kB/event (13%).