Adding automatic SUSY checks based on cvmfs file
The derivation framework needs to know if it's running on a sample that
has a DSID assigned to the SUSY group. We do that right now by
hard-coding a list of DSIDs. Instead, we can look up the DSIDs on cvmfs
(if it is available). Just takes a bit of string formatting.