TrigP1Test: Update Beamspot test
Re-implement the beamspot udpate test using a local sqlite file. Since
we lost the ability of using event modifiers in athenaHLT, it's now a
two-step process of first creating a dedicated bytestream file and then
running athenaHLT using a new python algorithm that takes care of
writing a new beamspot to the sqlite file at the appropriate time.