Some factorisation of the muon trigger decoding sequence
Moving slightly the decoding for the muon triggers to be within the individual menu sequences instead of within the algorithm sequences, and allowing to pass the RoI name to the decoding sequence. This will be needed for the late muon triggers (ATR-17553).
Also removed the 'eventAlg' setup of the decoding, as it doesn't seem to be used anywhere.