Move DB connection mgmt form Psc to HltEventLoopMgr
Move the management of DB connections before forking from the Psc to
HltEventLoopMgr (ATR-8907). This allows the sleep time to be configured
via the new `dbConnIdleWaitSec` property.

Moved the actual code to the `TrigRDBManager` helper class so we don't
have to change the friend relationship in `IRDBAccessSvc` in case we
need to move this code somewhere else again.