Ensuring MC event number is set in derivations
When derivations are made from EVNT files, we often encounter a
situation where the MC event number is not yet set, though the event
number is. This patch ensures that the MC event number in the xAOD event
info will be set to something sensible in such cases. No change, of
course, to the cases that were fine before.