TrigCOOLUpdateHelper: Configure the list of extensible folders
The TrigCOOLUpdateHelper now needs to be configured with the list of
"updateable" folders. That way it can make sure that those folders are
marked as "extensible" in the IOVDbSvc. This is checked during
configuration. In addition, if a folder update request is received for a
folder that is not configured as such, the application will fail. This
is to avoid that we are writing updates to COOL that are never applied
in the HLT.