Add configuration for merging EVNT_TR files
In ATLMCPROD-7665 it was noted that merging of EVNT_TR files in AthenaMP jobs fails.
This is due to the fact that the EVNT_TR input/output file arguments are of type
`argPOOLFile` and so try to use the default approach for merging reconstruction outputs.
This merge request creates a new `argEVNT_TRFile` argument type where a specialized
merging configuration is used. This should fix ATLASJT-407.