Zeroing MC event weights on old-style EventInfo
This solves the duplication of information in derivations between the
old-style EventInfo and the xAOD::EventInfo. The latter should be the
exclusive source for the information, and is used by all the PMG tools.
Therefore, we're clearing the event weights on the old-style EventInfo.
In release 22, we can simply outright drop the old-style EventInfo, but
in release 21 that will result in files that can't be read by athena (or
the derivation framework, for example). By putting this change into the
TP converter in 21.2, we get the change for all derivations immediately.