LumiCalc: merge website updates back into master
Several updates of iLumiCalc and associated scripts and html files
have been updated for the atlas-lumicalc website without ever being
committed back to the git master.  This patch brings the git master
back into sync with the code on the website, and provides bugfixes
for iLumiCalc that had been deployed on the website.