Add option to avoid overlap in DirectPhotonFilter slices
Currently the DirectPhotonFilter allows to specify a pTmin and pTmax which can be matched by any photon. It is then used to slice the gamma+jet sample into supposedly complementary slices separate by photon pT.

This is problematic e.g. if there are two photons in an event, the leading one with pT>pTmax and the subleading with pTmin<pT<pTmax. This event will be accepted in the given slice, but will also be part of a higher pT slice due to the leading photon.

Here, an option is to be introduced in the DirectPhotonFilter which allows to cut exactly on the leading (and/or subleading, ...) photon instead of inclusively.

Closes AGENE-1780