Removing additional photons from light mesons in Truth3 collection
The original cut on photons within TRUTH3 was that they not be
identified as coming from a pi0 and that they be at least 20 GeV. This
leaves a large number of photons in simple samples that are identified
as having come from a light meson and are quite low momentum. We
shouldn't need those either (etas and rhos are not interesting), so this
MR modifies the collection requirements to look particularly for those
not from light mesons either.

This saves us about 0.35 kB/event, or about 10% of the total file size.