Watching for PRW that has already been applied
In DAOD_PHYSLITE, we already have PRW applied at the derivation stage.
If tools set it up to run, it'll try to add a decoration -- but the one
that was added in the derivation is there and is const (since it was in
the file), so the code crashes. This adds a warning and early return in
such cases. I'm not explicitly checking for const decorations, just for
decorations that would have the same name as the ones to be added.