Our file limit should not be a warning
This warning message has been there for years, telling people that the
limit on the file size is 10 GB (now 15). This cannot possibly be a useful
warning. Either we need to check the appropriate condition, or we just
print this as info (as in this MR) and if the job dies then the error
message at that point should be made as meaningful as possible.