Adding large-R truth jet trimming if needed
Following the setup for EVNT files, this creates large-R truth jets as a
part of the standard truth content (aka TRUTH3). It's needed to fix some
of the crashes in the derivation framework from the latest nightly.
Those essentially come because separate methods are needed to setup
truth jet running on EVNT and AOD, and I hadn't tested enough on AODs.