Add di-tau to DAOD_PHYS
Hello,

This MR adds the low-pt and high-pt di-tau objects to DAOD_PHYS (ATLASG-1502).
High-pt di-taus are already part of the standard reconstruction, while low-pt di-tau reconstruction needs to be scheduled explicitly in R21 (it has been added to TauCommon).
We have added smart slimming lists for low- and high-pt dit-aus which include all Aux variables at the moment, to be refined soon.
The DAOD_PHYS structure had to be slightly reshuffled, because the low-pt ditau reconstruction relies on AntiKt2PV0TrackJets, so the the "Jet/MET" block needs to be scheduled before the low-pt di-tau reconstruction and track thinning.

Cheers,
Bertrand