WIP: modifications to permit skimming of DAOD_PHYS with derivation framework
This MR makes a number of alterations to the derivation framework and the job transforms executors to enable the skimming of DAOD_PHYS using the derivation framework machinery. The changes allow commands of this type:
'''
Reco_tf.py --inputDAOD_PHYSFile DAOD_PHYS.pool.root --outputD2AODFile test.pool.root --reductionConf SKIMTEST
'''
where SKIMTEST is a test format set up to test this workflow. The changes made are as follows:
- alterations to job transforms to add a new executor DAOD_PHYS->D2AOD_NAME, with NAME being the required skim-only format
- changes to job transforms framework to account for this new executor
- writing a length-one vector of Athena-readable event generator weights into DAOD_PHYS to stop the inputFilePeeker from complaining when it reads DAOD_PHYS

Formats that are to be skimmed from DAOD_PHYS will need to be set up or modified specifically - existing formats won't generally work as-is, as they use AOD-level information.