Use di-tau smart slimming instead of AllVariables
Hello,

As reported in !27694 , the introduction of a smart slimming list for DiTauJets and DiTauJetsLowPt collections has caused trouble for DAOD formats using `AllVariables` for these collections. The use of AllVariables is not allowed for large formats. The smart collections have to be used instead. As far as I can see, only DFHiggs needs that change. TAUP5 still uses AllVariables for DiTauJetsLowPt, but it's a small CP format which is not subject to the AllVariables restriction. I am running ART tests with the fix for HDBS1 and HIGG4D2, it seems fine.

Cheers,
Bertrand