Towards PFlow-seeded tau reconstruction: baby steps
Hello,

This MR adds a tauRec flag to specify the jet collection that is used to seed the tau reconstruction.
The proposed implementation is via a flag, rather than a property to be set on the TauProcessorTool, which seems more error prone.
A priori there won't be a need to run LC-jet-seeded and PFlow-jet-seeded tau reconstruction in the same job, it will be either one or the other.
So a TauProcessorTool property does not seem required.

Cheers,
Bertrand