Reducing log file noise from vertex finding
These tools put an incredible volume of noise into the logs. I just
can't believe that this is really "INFO" level output -- in the first
case (JetFitterRoutines) the same message is already at DEBUG level in
SeedNewtonTrkDistanceFinder. In the second case this provides no useful
information for the user, just that a candidate is being deleted, so it
can't possibly be something that a typical user should make use of.