Need ME q/p for bad muon decoration
For the time being, we need this ME q/p for bad muon decoration. MCP are
going to look into other ways to do the bad muon decoration (including
potentially doing it upstream in the derivation framework as Jet/MET do)
and will let us know if some variables can be dropped after that.