Changing link repoint to vector input
For a variety of reasons, truth links from a particle type may not point
to particles of the same type (e.g. photons faking electrons may have a
link from a reco electron to a true photon). We shouldn't try to save
all the hadrons that fake leptons and photons (that's for a CP
derivation), but for most applications if we save the particles anyway
we might as well make the correct link.

This MR extends the link repointer to look at multiple input collections
and resolve those cases.