Adding support for an alternative mass definition
The existing mass definitions do not include the way that the TCC
jet mass is stored.  This adds a new more direct mass definition
that works for TCC jets as well as any other jets where the final
mass definition is the four-vector's mass (not stored as a
separate moment).