AthenaServices: Define timestamp in dummy event context.
In AthenaEventLoopMgr when there is no event selector defined, we need to
create a dummy event context.  Previously, the EventID in this dummy context
defined run/event/lumi numbers, but not a timestamp.  This was observed
to cause problems in MuonAlignmentCondAlg, which accesses conditions during
initialize(), in tests that don't define an input file.  Define a dummy
timestamp as well to prevent this.