Re-enabled the generation jobReport.json in Event Service jobs
As a side effect of MR!25230 the generation of job report files was
disabled for the Event Service jobs. OTOH, if at the end of an Event
Service job the pilot does not find the report file, it produces an
error. This commit fixes this issue by re-enabling the generation
of job reports in AtlasG4_tf and Sim_tf, when they run Event Service