Protecting forward JVT call
There's no need to compute forward JVT if we don't need it. It is needed
in two places (in MET): 1) in the actual MET computation if fJVT is
true; 2) in the MET significance calculation if do pileup jets is true.
Therefore protecting the fJVT call with those two parameters.

This will mean that if there's someone out there not caring about either
of the above things but still wanting to use fJVT, they won't get it
decorated, but I'm having trouble imagining someone who wants to use
fJVT in their analysis but not in their MET (and I can't think of anyone
in SUSY who explicitly cuts on fJVT without using it in MET).