Updates to Pythia8 LHEF examples
The current LHEF example is meant to run as athena job options; adding a
note inside it to be explicit about that.

The include file uses the runArgs to make an input file and then
hard-codes the location of the input file; changing that to use the
runArgs as seems to be intended.

Adding a working example of LHEF usage with the transforms. This was
rather non-trivial, so I think it's useful.