Fixing more warnings in truth derivations
Two more warnings here. The first is to do with using out-dated event
densitity algorithms (fixed by moving to the new service). Second is
related to conditions reading, because some conditions files are added
by default to the conditions DB service. For the time being, simply
removing those files from the catelog in the case that we don't have a
conditions tag set as a part of the job.  That means no
autoconfiguration off of an AOD, and no explicit setting in the
transform. Should only be the case for truth derivation jobs, in which
case we are safe, as we don't need any conditions information anyway.