Made it possible to use VS Code with the WorkDir project.
In order for VS Code to parse our code correctly, it has to be told
explicitly how each source file needs to be compiled. (Otherwise it would
not find the headers from LCG/TDAQ/Gaudi to interpret our code.) This
update makes sure that when the user sets up the build of the
WorkDir project in the "usual layout", in a directory called build/
just beside the athena/ directory, VS Code would automatically find
the correct compile_commands.json file that it needs, when the user
opens the athena/ directory for editing.