AthContainers: Adjust overload ordering to work better with python.
pyroot doesn't follow C++ rules for overload resolution.  Instead, it tries
calling methods in order and accepts the first one that works.
This implies that if D derives from B and you you list foo(const B*)
followed by foo(D*), you'll only ever be able to call the first of these
from python even if you call it with a D instance.  This can be avoided
by reordering the method declarations.

This affects the setStore() methods.  Reorder them so that we can call
the non-const versions from python.