Updates for PHYSLITE running
This is the first set of updates to SUSYTools to allow running on
DAOD_PHYSLITE. The most invasive is that I've added an option to allow a
b-tagged jet collection key as an override. The remainder are mostly
options to set the collection names in the case that we're running on
PHYSLITE, and a few minor tweaks.