Fixing two issues with previous commits
Adding an "allow empty" option for string settings, in case the string
really can be left empty and we don't need a default value.

Fixing a stupid mistake in the jet code where I declared a variable
earlier and then had it re-declared in scope (that really should be a
compiler warning at some point...).