Updating Tile software to run on 12-bit ADC as well as on 10-bit ones.
The code is moved from an update that wasn't finished before ATLAS
opened its software in December 2018. The move to the public repository
was done via a patch file generated from the old repository.
The command git apply --reject patchfile.patch was used
together with a carefull inspection of the rejected hunks.