CaloRec: MT-safe version of AOD cell thinning.
Cells get thinned for AOD in an ad-hoc, thread-unsafe way:
A new cell container is made, AODCellContainer, with the cells
to be kept.  The links from clusters to cells are then modified
in place to point at the new container.

Add components to allow using the new MT-safe thinning.
Add CaloThinCellsBySamplingAlg and CaloThinCellsByClusterAlg
to request keeping cells in a given sampling or around clusters
in a given container.  A difference from the existing thinning
is that in the new case, we're really thinning the cell container,
so the cell container written to the AOD will still be called AllCalo.
To help existing code that may be using the name AODCellContainer,
also introduce CaloCellContainerAliasAlg which adds an alias
in this name.

Set up configuration functions for the new thinning.
New thinning components are not yet enabled, though.