Load the FastCaloSim V2 parametrization in a separate Service
When running in AthenaMT, we only want to load the parametrization once.
For compatibilty with AthenaMT running the simulators themselves have to be AlgTools
though, so it makes sense to load the parametrization in a separate Service
which will then only be loaded once.

Making this change in the 21.0 branch initially, to make it easier to sweep changes to the parameterization
between 21.0 and master in the future.