PHYSLITE : Adding tool to keep only the highest pT2 vertex
This adds a small tool to keep only the highest track sum pT2 vertex in
the primary vertex container. My paranoia requires that I check the
vertices for sum pT2, and that they're primary vertices, so that we're
sure we have the right one. All my tests show that it's the 0th item in
the list, as one would expect.

This tool is then added to PHYSLITE to reduce a few hundred bytes off.