TruthParticle: Adding extra test to parent/child
In some conditions a truth vertex (either production or decay) might be
saved in a different collection from the daughters of a particle, causing
the vertex to not know how many children or parents the particle has.
This is particularly possible in case we save vertices for only some
particles. An extra test lets us ensure that we avoid this case.