Change names of output PRD containers for Run 2 muon triggers
Changing the names of the output containers (and relevent input containers) for the RDO to PRD decoding for the Run 2 muon triggers. This is to avoid clashes in trying to write to the same containers when running the offline reconstruction and the legacy run 2 triggers at the same time.
Since the L2MuonSA config is used for both run 2 and run 3 triggers, this change only applies when we run in the run 2 mode (for doMT=False). The TrigMuonEF configuration is only used for the run 2 trigger case, so the changes are applied directly.
Fixes ATR-20615