PHYSLITE: Adding variables for running
Adding a bunch of variables for muon selection tools and MET making.
These are needed to get muons fully working in analysis. I hope that we
will be able to reduce the amount of information soon, with the help of
MCP and some modifications to their tools to run more decisions
upstream.