TrigMessageSvc: Implement asynchronous reporting
Implement asynchronous reporting during the event loop to avoid
potential deadlocks (e.g. ATR-20664). This is modelled after Gaudi's
InertMessageSvc using a separate thread for the message reporting.
However, since we do not want to have a separate thread around before
forking, we limit this feature to the event loop.