Repairing merging for truth derivations
We need certain metadata for reconstruction that (of course) does not
exist in a truth derivation. This MR looks to see if we are merging
truth derivations, and if we are, it disables some of the problematic
items, and sets up the things we neeed. It is largely taken from
PATJobTransforms/skeleton.AODtoDAOD_tf.py