PHYSLITE updated trigger matching
We were missing a rather important line for PHYSLITE to get the trigger
matching containers included in the derivation output. Now they should
be there.

Also adding a warning to the trigger matching tool when it
(automatically) removes a trigger that the user has requested and it has
decided should not be included.