Backported the build script improvements for AthAnalysis and AthDerivation.
In order to make the nightlies a bit more resilient in case of some external build
failure, I backported an update to the build_externals.sh scripts of
AthAnalysis and AthDerivation.

Note that I didn't touch any of the other projects. For the non-Gaudi/Athena
projects this update is irrelevant. And for the other Gaudi/Athena based ones
I just didn't bother. Since we're not building them in this branch anyway.
(And they have been updated correctly in master already, where the
scripts look a little different anyway.)