TrigPSC: Define printfunc before including JO files.
For compatibility between py2 and py3, athena JO files use `printfunc()'
instead of either the py2 print statement of the py3 builtin print()
function.  (The reason for not just using print() along with
from __future__ import print_function is that with an included file,
the print_function future can't be limited to one file.)

Define printfunc here to be compatibile with what's done in athena.