Add feature access support methods for the new MT TDT feature access
In order to migrate the ID TRigger feature access, we need methods to wrap the new
MT feature access to minimise actual disruption to the existing code.

This adds the feature access methods in the base class that adds the tracks to the
TeackSelector for subsequent use.

These changes are needed before subsequent development can take place.