Apparent (critical) bug in Pythia8_i
This fixes a critical bug in Pythia8_i that results in a memory
allocation out of control and the job being killed.

The condition comes up when LHE3 weights are enabled and Pythia8 does
not have any shower weights to provide. In that case, this loop begins
at 1 and the stopping condition is that the iterator (being increased)
is equal to zero, which won't happen for a very long time (the int has
to flip over and reach zero again). This has caused a number of test
failures already.