Updating tau rec tools to do less in beginEvent
BuildTruthTaus was previously using beginEvent as an opportunity to look
at metadata -- to do a sort of late configuration on the fly. That isn't
an idea way to do things here, so changing to use a simple function and
a member variable cache, which is called in any of the applicable
functions when the value is actually needed.

This solves a number of errors when running derivations in athenaMP.