Adding c*L offset to stopped particle decays
When we say in stopped particle land that a particle decays at 10 ns, we
don't really mean 10ns after the bunch crossing, but at 10ns plus the
distance to the IP times the speed of light after the bunch crossing (ie
out of time by 10 ns). This MR adds an option (default to true) to
include this offset in the decay time.