PHYSLITE: Disabling fJVT tool and selection
We now have a tool running in the derivation to calculate the PFlow jet
FJVT and decorate it according to Jet/MET group recommendations.
Therefore, this MR disables the fJVT algorithm from the analysis
algorithm package and the subsequent selection using it.

It's perhaps worth pointing out that the decoration in the derivation
framework seems to use the JetForwardPFlowJvtTool tool:
https://gitlab.cern.ch/atlas/athena/blob/21.2/PhysicsAnalysis/DerivationFramework/DerivationFrameworkJetEtMiss/python/ExtendedJetCommon.py#L431

while the analysis recommendations:
https://twiki.cern.ch/twiki/bin/view/AtlasProtected/PileupJetRecommendations#fJVT

seem to suggest that the JetForwardJvtTool tool can be used for either
EMTopo or PFlow jets. The analysis algorithms currently follow the
latter setup (just the JetForwardJvtTool for all); it's not clear to me
if this is really the recommendation or is just an inconsistency that
needs to be fixed. Mercifully, since we are using the derivation
decoration, it doesn't matter for PHYSLITE.